Refactor stop/delete logic into pkg/instance
